requirements.txt
asgiref==3.5.0
Django==3.2.13
django-filter==21.1
djangorestframework==3.13.1
pandas==1.3.5
pytz==2022.1
regex==2022.3.15
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
waitress==2.1.1
whitenoise==6.0.0
